add Readme.md inside test_workspace folder and ad a filter containing that filename to cleanup funtion

+++ tests/loader_test.py
@@ -94,6 +94,8 @@ class DxxXmlValidatorTest(unittest.TestCase):
 
     def cleanup_workspace(self, workspace):
         for f in os.listdir(workspace):
+            if "Readme.md" in f:
+                continue
             os.remove(os.path.join(workspace, f))
